[new-channels-ui] Allow creation and update of QST Server channels
In order to update a QST Server channel, a new password is assigned every time a new ticket_code is use. By using this approach, the "password" field in the local gateway app is configured automatically, and the user don't have to enter the new password manually. 
In addition, password and password_confirmation fields were explicitly removed from the form of the QST Server Channel, which supports the approach of an automatic configuration of these fields.