Cleanup expired Let's Encrypt CA
Let's Encrypt old Certificate Authority expired on Sept 30, 2021.
Most of InSTEDD's infra uses those certificates, so this patch
makes new certificates work again.

I wasn't able to pin-point which specific packages needed to be updated
in order for the certificates to work, so I've just simply upgraded
every package available. We need to upgrade the project's Ruby version
and base dependencies anyway, so I don't want to keep investing effort
on an unsupported platform.

instedd/docker-nginx-rails#2 was needed to be able to run the tests in
the CI using docker-compose